Add sequence that breaks up a collection
Add a sequence structure that takes a collection and a span, and vends
sub-sequences of the collection, each with a maximum count of the given
span.  The last element will be shorter if the collection’s count isn’t
divided equally by the span.

The estimate of the sub-collection count is exact for random-access
collections and speed-optimized for wimpier collections.