Add sequence that clusters elements from another
Add a sequence structure that takes another sequence and a span, and
vends arrays of the inner sequence’s elements, each with a maximum
count of the given span.  The last element will be shorter if the inner
sequence’s element count isn’t divided equally by the span.

The estimate of the array count is exact for random-access collections
and inexact for wimpier sequences.

Clarify the name of an earlier test.