Add collection that chunks another
Add a collection structure that takes another collection and a span,
and uses sequential sub-sequences of the inner collection as elements,
each with a maximum count of the given span.  The last element will be
shorter if the inner collection’s element count isn’t divided equally
by the span.

Include a lazy generator for chunked collections.